requirements.txt
boto3==1.9.140
botocore==1.12.140
Django==2.0.1
django-ckeditor==5.7.1
django-crispy-forms==1.7.2
django-js-asset==1.2.2
django-storages==1.7.1
docutils==0.14
jmespath==0.9.4
jsmin==2.2.2
Pillow==5.0.0
pkg-resources==0.0.0
psycopg2==2.8.3
python-dateutil==2.8.0
pytz==2017.3
s3transfer==0.2.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.2
